Fix exit_status example
In the example in the documentation that calls exit_status and in the
smoke integration test exit_status is called before the channel is
closed and it will therefore always return Ok(0).

I have fixed this by calling wait_close() first and by adding a
bad_smoke test that calls false and checks that the exit_status is 1.