remove SessionBinding trait
I'd gradually removed the usage of it while I was hacking through
with my lifetime changes.  I suppose it could be added back
but I confess that I removed it because I found it confusing
and it didn't look like it would fit on my first pass through
the code.