Add script for running integration tests
This makes it possible to run the integration tests without
requiring that the user change their local ssh configuration.
This is desirable because some sites have strict controls over
the local ssh configuration files.

This commit adds a script that spawns a local copy of the ssh
daemon running on an alternate port with a specific configuration
that is known to successfully pass the test suite.

This has only been tested on my local macos machine so far.