Make channel tests more robust to stderr output
At the start of feb 2019 the tests started to fail on macos on Travis.  No
changes were made to the repo, so the problem must be environmental.

It turns out that the server was emitting stderr output, and since the
tests were not consuming stderr, this made `channel.eof()` always
return false.

The resolution is to drain both stdout and stderr in the tests before
making assertions with `channel.eof()`.

I've also updated the doc comment on `channel.eof()` to reflect this,
as it is a bit of a sharp edge.