Expose more known_hosts related constants
This brings us up to date with the current set of host key and known
host key enum variants so that we can parse and edit current known_hosts
files.

This also adds a convenience `From` impl that allows converting from
the host key type to a known hosts entry type.