improve handshake error reporting on windows
I'm hitting a kex protocol error when talking from windows->ubuntu 19
and that wasn't getting reported as an error on the client side
because the `rc` function only emits an error if one has been recorded
against the session object, despite being reported via the parameter.

I think it's worth revisiting this more broadly, but for now, let's
report this error in the handshake case.