move tcpstream assignment to its own function
The recent move to take ownership of TcpStream exposed an issue with
the `handshake` method: if the stream is non-blocking then it may
take several attempts to handshake, but only the first one is able
to transfer ownership.

My initial thought was just to make the TcpStream a required parameter
to `new`, but we have some tests that work with known hosts and the
ssh agent that don't require a tcpstream.

I'm going to review those and see if there is a cleaner overall
solution, but that will likely require more substantial API changes.

For now, the simplest change is to add a separate `set_tcp_stream`
function to make the stream ownership transfer explicit and distinct
from the handshake.

Refs: https://github.com/alexcrichton/ssh2-rs/issues/17