tighten up `Session::rc`
Previously, this function could return None even if passed an error
result.  This could happen in situations where no error had been
recorded against the session object, but where libssh2 had returned
an error condition anyway.

It was also possible for the prior implementation to return the
wrong information in a similar situation: if an error had been
recorded against the session, that error would be returned rather
than the current error code.

This commit tidies up the implementation of this function by
always returning an `Error` object that corresponds to the
provided `rc` parameter.  If the last error message recorded
against the `Session` has a matching error value then that
error message will be used, otherwise we'll synthesize one
just from the error code.