add shutdown and close methods
This commit exposes the `shutdown` method of `Sft` and `close` method of
`File` to free resources. Previously this was done synchronously and
therefor not usable in async wrappers.

Whith the methods exposed they can be polled until success. If
successful a flag is set to notify the destructor that no resources
need to be freed anymore.

However, if resources were not cleaned up by calling `close` or
`shutdown` the destructor will clean up synchronously.