Precise error codes when using the SFTP subsystem
When an error happens while using the SFTP subsystem, the error returned
by libssh2 is LIBSSH2_ERROR_SFTP_PROTOCOL, which only indicate the
origin of the error, but not its nature. In order to know what caused
the error, libssh2_sftp_last_error() needs to be called, and would
return one of the LIBSSH2_FX_* error codes.

Previously, the libssh2 crate didn't call libssh2_sftp_last_error() when
LIBSSH2_ERROR_SFTP_PROTOCOL was reported, but simply returned an obscure
Error. This is now fixed.

Fixes #182