Add tcp_stream() function to obtain raw fd/socket
This adds the convenient methods to obtain the raw file descriptor (Unix) or socket (Windows)
which can be used lifted to the TcpStream which in turn can be used to
get some useful information.

This method was mentioned in the documentation to `set_tcp_stream()` but
was not provided at this moment. It must be fixed now.