Expose more SFTP file types (#213)
As a user of the `ssh2` crate it was not possible to figure out if a
file was a named pipe, a character device, a block device, or a socket.

By changing FileType to an enum we expose more information about the
file type when it's available.

I tried to keep the API as backwards compatible as I could. Not all bit
patterns of `perm` lead to a valid/known filetype, so we end up with an
`Other` variant. Another design would have been to make
`FileStat::file_type` return an `Option<FileType>` but that would have
been a breaking API change.