Haiku: always link against shared library when cross-compiling
This is a niche change for the particular scenario where libssh2-sys is built as part of the rustc build. Due to various reasons, this is done with a cross-compiler. When cargo is built as part of the extended build, it also links against the curl shared library, which on Haiku links against libssh2.so. When libssh2-sys includes a static instance of all the symbols, this leads to a linker error because libcurl.so points to hidden symbols.

This change in the build-script short-circuits this problem by always falling back on the system library, specifically in this case.