libssh2-sys: Fix bug in sftp_readdir where files could be fixed
This adds a commit that fixes a bug in sftp_readdir where if the inital
buffer we supply is not big enough for a file name, that file would be
skipped.

We mitigated this issue by using a larger buffer to begin with, but
that is a bandaid.

See https://github.com/libssh2/libssh2/issues/714.

Fixes #217