Pass `comment` as `CString`
The `libssh2_knownhost_addc()` expects `comment` to be a C string: it
copies `commentlen + 1` bytes internally from the buffer pointed by
`comment` pointer. Granted, this is misleading: one would expect that
if a function takes a pointer and length of data, it wouldn't access
memory past this limit.

Passing a `&str` as a pointer and its length is incorrect and results
in segmentation fault with `rustc` > 1.78.0 on macOS 14.6.1 (23G93)
and FreeBSD 14 - we pass empty &str literal to the function.